requirements.txt
numpy==1.26.4
pandas==2.1.4
scipy==1.11.4
scikit-learn==1.7.2
matplotlib==3.8.0
openpyxl==3.1.5
SimpleITK==2.5.2
pyradiomics==3.0.1
xgboost==2.0.3
joblib==1.2.0
